+++ packages/extension-base/src/background/handlers/helpers.ts
@@ -1,4 +1,4 @@
-// Copyright 2019-2023 @polkadot/extension authors & contributors
+// Copyright 2019-2023 @polkadot/extension-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 export function withErrorLog (fn: () => unknown): void {
