+++ packages/extension-base/src/background/handlers/subscriptions.ts
@@ -1,4 +1,4 @@
-// Copyright 2019-2023 @polkadot/extension authors & contributors
+// Copyright 2019-2023 @polkadot/extension-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import type { MessageTypesWithSubscriptions, SubscriptionMessageTypes } from '../types.js';
