+++ packages/extension-base/src/background/index.ts
@@ -1,4 +1,4 @@
 // Copyright 2019-2023 @polkadot/extension-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export { canDerive } from './canDerive.js';
+export { default as handlers, withErrorLog } from './handlers/index.js';
