+++ packages/extension-ui/src/Popup/AuthManagement/index.tsx
@@ -1,7 +1,7 @@
 // Copyright 2019-2023 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { AuthUrlInfo, AuthUrls } from '@polkadot/extension-base/background/handlers/State';
+import type { AuthUrlInfo, AuthUrls } from '@polkadot/extension-base/background/types';
 import type { ThemeProps } from '../../types.js';
 
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
