+++ tsconfig.base.json
@@ -8,8 +8,8 @@
       "@polkadot/extension": ["extension/src/index.ts"],
       "@polkadot/extension/*": ["extension/src/*.ts"],
       "@polkadot/extension-base": ["extension-base/src/index.ts"],
-      "@polkadot/extension-base/background/handlers": ["extension-base/src/background/handlers/index.ts"],
-      "@polkadot/extension-base/packageInfo": ["extension-base/src/packageInfo.ts"],
+      "@polkadot/extension-base/background": ["extension-base/src/background/index.ts"],
+      "@polkadot/extension-base/background/types": ["extension-base/src/background/types.ts"],
       "@polkadot/extension-base/page": ["extension-base/src/page/index.ts"],
       "@polkadot/extension-base/stores": ["extension-base/src/stores/index.ts"],
       "@polkadot/extension-base/utils": ["extension-base/src/utils/index.ts"],
